Rutgers CS Schedule Checker
This project utilizes Spring Boot to create a REST API that allows users to download a file with information on Rutgers Computer Science courses they are eligible to take or a sample schedule for a target course. The front-end is developed with Vanilla Javascript while the back-end uses Node JS and Java.