Rutgers CS Schedule Checker
This project utilizes Spring Boot to create a REST API that allows users to download a file with information on eligible Rutgers Computer Science classes or a sample schedule for a target course based on their input. The front-end is developed with Vanilla JavaScript, while the back-end uses Node.js and Java.